requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
python-dateutil==2.8.2
pyyaml==6.0.1
cryptography==42.0.5
colorama==0.4.6